fix migration rollback

+++ database/migrations/2024_05_11_203010_add_extra_information_on_variants.php
@@ -20,8 +20,10 @@ return new class extends Migration {
      */
     public function down(): void {
         //
-        Schema::table('pet_variants', function (Blueprint $table) {
-            $table->dropColumn('description');
-        });
+        if (Schema::hasTable('pet_variants')) {
+            Schema::table('pet_variants', function (Blueprint $table) {
+                $table->dropColumn('description');
+            });
+        }
     }
 };
